fix: typings and general cleanup
The intention of these changes was to add the `"types"` field to the
package.json` file to resolve #9.
However, I've also included general improvements to the structure of the
project. Namely dropping the npm workspaces setup in favor of a more
traditional repository structure. This should make it easier for
contributors to understand the project and make changes.

- Add `"name"` fields to jobs for GitHub Actions
- Add `"types"` field to `package.json` to resolve #9
- Add `// @ts-check` to `index.test.js` for static type checking (since
  it is not included in the `tsconfig.json`)
- Add configuration for "Buy Me a Coffee" sponsor button (see
  [the GitHub docs](https://docs.github.com/en/repositories/managing-your-repositorys-settings-and-features/customizing-your-repository/displaying-a-sponsor-button-in-your-repository)
  for more details on what this is and how to set it up)
- Bump `"node-version"` to `18` for GitHub Actions. This should reduce
  issues as we can internally target Node 18
- Bump package version to `"2.1.0"`
- Remove abbreviations in commands throughout for improved human
  readability/comprehension
- Remove `.npmignore` in favor of a concise `"files"` field in
  `package.json`
- Remove configuration for `*.js` files in `.editorconfig` to
  standardize 2-space indentation
- Remove `CHANGES.md` from `.gitignore` (this was something I used to
  use in my projects before discovering a better workflow for commit
  messages)
- Remove workspaces configuration from `package.json` in favor of a more
  traditional repository structure
- Update `LICENSE` to `LICENSE.md` from
  [IQAndreas/markdown-licenses](https://github.com/IQAndreas/markdown-licenses)
  for improved readability
- Update format of HTML in `README.md` for improved readability
- Update metadata (shields.io badges, etc.) in `README.md` to be
  consistent. Also adds a badge for license, version, gzip size, last
  commit, and buymeacoffee (removing the buymeacoffe button and nodei.co
  info)
- Update order of `package.json` fields to match the order in the
  [npm documentation](https://docs.npmjs.com/cli/configuring-npm/package-json)
- Update `package.json` to use abbreviated major version syntax (e.g.
  `"^1"` instead of `"^1.x.x"`)
