+++ .github/CODEOWNERS
@@ -1,0 +1,2 @@
+# learn more: https://docs.github.com/en/repositories/managing-your-repositorys-settings-and-features/customizing-your-repository/about-code-owners
+* @WillTDA
