+++ .github/FUNDING.yml
@@ -1,0 +1,2 @@
+# learn more: https://docs.github.com/en/repositories/managing-your-repositorys-settings-and-features/customizing-your-repository/displaying-a-sponsor-button-in-your-repository
+buy_me_a_coffee: willtda
