--- tests/package.json
@@ -1,12 +1,0 @@
-{
-  "name": "tests",
-  "version": "0.0.0",
-  "scripts": {
-    "test": "mocha **/*.test.js"
-  },
-  "devDependencies": {
-    "@types/mocha": "^10.0.1",
-    "mocha": "^10.2.0"
-  },
-  "private": true
-}
